+++ assets/src/scripts/monitoring-location/components/hydrograph/selectors/parameter-data.js
@@ -1,3 +1,4 @@
+import {DateTime} from 'luxon';
 import {createSelector} from 'reselect';
 
 import config from 'ui/config';
@@ -29,6 +30,23 @@ export const getAvailableParameterCodes = createSelector(
                 const parameterCode = variable.variableCode.value;
                 const measuredParameterCode = parameterCode.replace(config.CALCULATED_TEMPERATURE_VARIABLE_CODE, '');
                 const isUVParameterCode = config.uvPeriodOfRecord && measuredParameterCode in config.uvPeriodOfRecord;
+                const isGWParameterCode = config.gwPeriodOfRecord && measuredParameterCode in config.gwPeriodOfRecord;
+                const uvPeriodOfRecord = isUVParameterCode ? config.uvPeriodOfRecord[measuredParameterCode] : null;
+                const gwPeriodOfRecord = isGWParameterCode ? config.gwPeriodOfRecord[measuredParameterCode] : null;
+                let periodOfRecord;
+                if (!uvPeriodOfRecord) {
+                    periodOfRecord = gwPeriodOfRecord;
+                } else if (!gwPeriodOfRecord) {
+                    periodOfRecord = uvPeriodOfRecord;
+                } else {
+                    periodOfRecord = {
+                        begin_date: DateTime.fromISO(uvPeriodOfRecord.begin_date) < DateTime.fromISO(gwPeriodOfRecord.begin_date) ?
+                            uvPeriodOfRecord.begin_date : gwPeriodOfRecord.begin_date,
+                        end_date: DateTime.fromISO(uvPeriodOfRecord.end_date) > DateTime.fromISO(gwPeriodOfRecord.end_date) ?
+                            uvPeriodOfRecord.end_date : gwPeriodOfRecord.end_date
+                    };
+                }
+
                 const hasWaterAlert = !!(isUVParameterCode && config.WATER_ALERT_PARAMETER_CODES.includes(measuredParameterCode));
                 let waterAlertDisplayText;
                 let waterAlertTooltipText;
@@ -56,8 +74,7 @@ export const getAvailableParameterCodes = createSelector(
                     timeSeriesCount: seriesList.filter(ts => {
                         return ts.tsKey === 'current:P7D' && ts.variable === variable.oid;
                     }).length,
-                    periodOfRecord: config.uvPeriodOfRecord && measuredParameterCode in config.uvPeriodOfRecord ?
-                        config.uvPeriodOfRecord[measuredParameterCode] : null,
+                    periodOfRecord: periodOfRecord,
                     waterAlert: {
                         hasWaterAlert,
                         subscriptionParameterCode: hasWaterAlert ? measuredParameterCode : '',